[IMP] Add server version management and auto download
Removed the lsp and refactored the config into a lua plugin with
modules which configures and enables the lsp on `setup()`.

The lua modules handle fetching the latest version from the github
repo, checking whether there is a newer version than the one installed
and expose an :OdooLsInstall function for users to install the newest
ls server on demand.